Move noise clamp tests to a separate file
