+++ projects/rocprofiler-compute/tests/test_utils.py
@@ -7468,252 +7468,6 @@ def test_validate_roofline_csv_invalid_inconsistent_columns():
 
 
 # =============================================================================
-# TESTS FOR NOISE_CLAMP: Multi-Pass Profiling Variance Handling
-# =============================================================================
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_clamping_behavior():
-    """Core behavior: positives unchanged, negatives clamped to 0."""
-    import numpy as np
-
-    from utils.parser import to_noise_clamp
-
-    # Scalar: positive unchanged
-    assert to_noise_clamp(1000.0, 100000.0) == 1000.0
-    # Scalar: negative clamped
-    assert to_noise_clamp(-100.0, 1000000.0) == 0.0
-
-    # Series: mixed values
-    diff = pd.Series([100.0, -50.0, 200.0, -100.0])
-    ref = pd.Series([1e6, 1e6, 1e6, 1e6])
-    result = to_noise_clamp(diff, ref)
-    pd.testing.assert_series_equal(result, pd.Series([100.0, 0.0, 200.0, 0.0]))
-
-    # NumPy array
-    diff_np = np.array([100.0, -50.0])
-    ref_np = np.array([1e6, 1e6])
-    result_np = to_noise_clamp(diff_np, ref_np)
-    np.testing.assert_array_equal(result_np, np.array([100.0, 0.0]))
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_zero_reference():
-    """Edge case: zero reference should not cause division by zero."""
-    from utils.parser import to_noise_clamp
-
-    assert to_noise_clamp(-100.0, 0.0) == 0.0
-    result = to_noise_clamp(pd.Series([-100.0]), pd.Series([0.0]))
-    assert result.iloc[0] == 0.0
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_warning_above_threshold():
-    """Warning recorded when relative error >= 1%."""
-    from utils.parser import (
-        clear_noise_clamp_warnings,
-        get_noise_clamp_warnings,
-        to_noise_clamp,
-    )
-
-    clear_noise_clamp_warnings()
-
-    # 2% error (above 1% threshold) - should record
-    to_noise_clamp(pd.Series([-20000.0]), pd.Series([1000000.0]))
-
-    stats = get_noise_clamp_warnings()
-    assert stats["count"] == 1
-    assert stats["max_rel"] >= 0.01
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_no_warning_below_threshold():
-    """No warning when relative error < 1%."""
-    from utils.parser import (
-        clear_noise_clamp_warnings,
-        get_noise_clamp_warnings,
-        to_noise_clamp,
-    )
-
-    clear_noise_clamp_warnings()
-
-    # 0.5% error (below 1% threshold) - still clamped, no warning
-    result = to_noise_clamp(pd.Series([-5000.0]), pd.Series([1000000.0]))
-    assert result.iloc[0] == 0.0
-    assert get_noise_clamp_warnings()["count"] == 0
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_empty_input():
-    """Empty inputs should return empty without error."""
-    from utils.parser import to_noise_clamp
-
-    result = to_noise_clamp(pd.Series([], dtype=float), pd.Series([], dtype=float))
-    assert len(result) == 0
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_threshold_boundary():
-    """Exactly 1% error should trigger warning (>= not >)."""
-    from utils.parser import (
-        clear_noise_clamp_warnings,
-        get_noise_clamp_warnings,
-        to_noise_clamp,
-    )
-
-    clear_noise_clamp_warnings()
-
-    # Exactly 1% error: -10000 / 1000000 = 0.01
-    to_noise_clamp(pd.Series([-10000.0]), pd.Series([1000000.0]))
-    assert get_noise_clamp_warnings()["count"] == 1
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamper_instance_isolation():
-    """Separate NoiseClamper instances should have independent state."""
-    import numpy as np
-
-    from utils.parser import NoiseClamper
-
-    clamper1 = NoiseClamper()
-    clamper2 = NoiseClamper()
-
-    clamper1.clamp(pd.Series([-20000.0]), pd.Series([1000000.0]))
-
-    assert clamper1.get_stats()["count"] == 1
-    assert clamper2.get_stats()["count"] == 0
-
-    clamper1.clear()
-    assert clamper1.get_stats()["count"] == 0
-    assert clamper2.get_stats()["count"] == 0
-
-    clamper1.clamp(np.array([-50000.0]), np.array([1000000.0]))
-    clamper2.clamp(np.array([-30000.0, -40000.0]), np.array([1000000.0, 1000000.0]))
-
-    assert clamper1.get_stats()["count"] == 1
-    assert clamper2.get_stats()["count"] == 2
-
-
-# =============================================================================
-# TESTS FOR NOISE_CLAMP: Upper-Bound (Max) Clamping
-# =============================================================================
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_scalar_exceeds_reference():
-    """Scalar values exceeding reference are clamped to reference."""
-    from utils.parser import to_noise_clamp
-
-    assert to_noise_clamp(150.0, 100.0) == 100.0
-    assert to_noise_clamp(100.0, 100.0) == 100.0
-    assert to_noise_clamp(99.99, 100.0) == 99.99
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_series_exceeds_reference():
-    """Series values exceeding their per-element reference are clamped."""
-    from utils.parser import to_noise_clamp
-
-    diff = pd.Series([50.0, 150.0, 100.0, 200.0])
-    ref = pd.Series([100.0, 100.0, 100.0, 100.0])
-    result = to_noise_clamp(diff, ref)
-    pd.testing.assert_series_equal(result, pd.Series([50.0, 100.0, 100.0, 100.0]))
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_ndarray_exceeds_reference():
-    """NumPy array values exceeding reference are clamped."""
-    import numpy as np
-
-    from utils.parser import to_noise_clamp
-
-    diff = np.array([50.0, 150.0, 200.0])
-    ref = np.array([100.0, 100.0, 100.0])
-    result = to_noise_clamp(diff, ref)
-    np.testing.assert_array_equal(result, np.array([50.0, 100.0, 100.0]))
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_scalar_reference_with_array():
-    """Scalar reference broadcasts correctly against an array difference."""
-    from utils.parser import to_noise_clamp
-
-    diff = pd.Series([50.0, 110.0, 99.0])
-    result = to_noise_clamp(diff, 100.0)
-    pd.testing.assert_series_equal(result, pd.Series([50.0, 100.0, 99.0]))
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_mixed_negative_and_exceeding():
-    """Both min-clamp and max-clamp fire in a single call."""
-    from utils.parser import to_noise_clamp
-
-    diff = pd.Series([-10.0, 50.0, 150.0, 100.0])
-    ref = pd.Series([100.0, 100.0, 100.0, 100.0])
-    result = to_noise_clamp(diff, ref)
-    pd.testing.assert_series_equal(result, pd.Series([0.0, 50.0, 100.0, 100.0]))
-
-    assert to_noise_clamp(100 - 105, 100) == 0.0
-    assert to_noise_clamp(100 - 80, 100) == 20.0
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_upper_warning_above_threshold():
-    """Warning recorded when upper-bound excess >= 1%."""
-    from utils.parser import (
-        clear_noise_clamp_warnings,
-        get_noise_clamp_warnings,
-        to_noise_clamp,
-    )
-
-    clear_noise_clamp_warnings()
-
-    to_noise_clamp(pd.Series([105.0]), pd.Series([100.0]))
-
-    stats = get_noise_clamp_warnings()
-    assert stats["count"] == 1
-    assert stats["max_rel"] >= 0.01
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_upper_no_warning_below_threshold():
-    """No warning when upper-bound excess < 1%, but value is still clamped."""
-    from utils.parser import (
-        clear_noise_clamp_warnings,
-        get_noise_clamp_warnings,
-        to_noise_clamp,
-    )
-
-    clear_noise_clamp_warnings()
-
-    result = to_noise_clamp(pd.Series([100.005]), pd.Series([100.0]))
-    assert result.iloc[0] == 100.0
-    assert get_noise_clamp_warnings()["count"] == 0
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_nan_with_exceeding_values():
-    """NaN and None inputs return np.nan regardless of reference."""
-    import numpy as np
-
-    from utils.parser import to_noise_clamp
-
-    assert np.isnan(to_noise_clamp(None, 100.0))
-    assert np.isnan(to_noise_clamp(float("nan"), 100.0))
-
-
-@pytest.mark.noise_clamp
-def test_noise_clamp_per_element_reference_array():
-    """Each element is clamped against its own reference value."""
-    from utils.parser import to_noise_clamp
-
-    diff = pd.Series([60.0, 200.0, 80.0])
-    ref = pd.Series([100.0, 150.0, 50.0])
-    result = to_noise_clamp(diff, ref)
-    pd.testing.assert_series_equal(result, pd.Series([60.0, 150.0, 50.0]))
-
-
-# =============================================================================
 # Experimental Feature Tests
 # =============================================================================
